refactor: move landing page's specific component (from compoenent/home to modules/home)
