+++ src/modules/home.module.tsx
@@ -2,13 +2,13 @@ import React from "react"
 import { makeStyles } from "@material-ui/core/styles"
 import { CoverComponent } from "../core/components/cover.component"
 
-import HomeTitle from "../core/components/home/home-title.component"
-import HomeLarngearDescription from "../core/components/home/home-larngear-description.component"
-import HomeQualification from "../core/components/home/home-qualification.component"
-import HomeFooter from "../core/components/home/home-footer.component"
-import HomeTimeline from "../core/components/home/home-timeline.component"
-import HomeCarousal from "../core/components/home/home-carousel.component"
-import HomeAnnouce from "../core/components/home/home-annouce.component"
+import HomeTitle from "./home/home-title.module"
+import HomeLarngearDescription from "./home/home-description.module"
+import HomeQualification from "./home/home-qualification.module"
+import HomeFooter from "./home/home-footer.module"
+import HomeTimeline from "./home/home-timeline.module"
+import HomeCarousal from "./home/home-carousel.module"
+import HomeAnnouce from "./home/home-annouce.module"
 
 import landing1 from "../assets/images/background/landing-1.svg"
 import landing2 from "../assets/images/background/landing-2.svg"
