+++ src/modules/home/home-carousel.module.tsx
